move around docs components
