+++ src/docs/components/index.ts
@@ -1,4 +1,6 @@
 
+export * from './layout';
+export * from './markdown';
 
-export { default as PropsTable } from './PropsTable.svelte';
+export { default as PropsTable } from './props-table.svelte';
 export { default as Search } from './search.svelte';
