+++ src/docs/components/layout/index.ts
@@ -1,0 +1,7 @@
+export { default as BottomNav } from './bottom-nav.svelte';
+export { default as HamburgerMenu } from './hamburger-menu.svelte';
+export { default as NavBar } from './nav-bar.svelte';
+export { default as Sidebar } from './sidebar.svelte';
+export { default as SidebarGroup } from './sidebar-group.svelte';
+
+export * from './types';
