+++ src/docs/components/markdown/layout.svelte
@@ -1,6 +1,6 @@
 <!-- MDsveX Layout Component -->
 <script context="module">
-	import { a, h1, h2, h3, h4, h5, h6, p, ul } from '$docs/components/markdown';
+	import { a, h1, h2, h3, h4, h5, h6, p, ul } from '$components';
 
 	export { a, h1, h2, h3, h4, h5, h6, p, ul };
 </script>
