+++ src/docs/components/props-table.svelte
