--- src/docs/layout/index.ts
@@ -1,5 +1,0 @@
-export { default as BottomNav } from './BottomNav.svelte';
-export { default as HamburgerMenu } from './HamburgerMenu.svelte';
-export { default as NavBar } from './NavBar.svelte';
-export { default as Sidebar } from './Sidebar.svelte';
-export { default as SideBarGroup } from './SidebarGroup.svelte';
